+++ src/main/scala/org/dia/algorithms/mcc/MCCNode.scala
@@ -19,7 +19,7 @@ package org.dia.algorithms.mcc
 
 import scala.collection.mutable
 
-class MCCNode(var frameNum :Int, var cloudElemNum : Float) extends Serializable {
+class MCCNode(var frameNum :Int, var cloudElemNum : Double) extends Serializable {
 
   override def toString = s"($frameNum,$cloudElemNum)"
 
@@ -37,13 +37,13 @@ class MCCNode(var frameNum :Int, var cloudElemNum : Float) extends Serializable 
   var inEdges : mutable.HashSet[MCCEdge] = new mutable.HashSet[MCCEdge]
   var outEdges : mutable.HashSet[MCCEdge] = new mutable.HashSet[MCCEdge]
 
-  def connectTo(destNode: MCCNode, weight: Float): MCCEdge = {
+  def connectTo(destNode: MCCNode, weight: Double): MCCEdge = {
     val edge = new MCCEdge(this, destNode, weight)
     addOutgoingEdge(edge)
     return edge
   }
 
-  def connectFrom(srcNode: MCCNode, weight: Float): MCCEdge = {
+  def connectFrom(srcNode: MCCNode, weight: Double): MCCEdge = {
     val edge = new MCCEdge(this, srcNode, weight)
     addIncomingEdge(edge)
     return edge
@@ -61,7 +61,7 @@ class MCCNode(var frameNum :Int, var cloudElemNum : Float) extends Serializable 
     frameNum
   }
 
-  def getCloudElemNum : Float = {
+  def getCloudElemNum : Double = {
     cloudElemNum
   }
